refactor(bus): flip bus files' Message/MessageKind imports to super (P4b/3)
Mechanical: bus production files now reference the canonical types via
super:: (crate::bus) instead of crate::microsvc::. Remaining crate::microsvc
references in bus are error.rs (B4), knative ingress (B5), and Service-based
test modules (B6).

Implements [[tasks/bus-decomposition-phase1]]
