+++ src/bus/nats_bus.rs
@@ -29,7 +29,7 @@ use super::nats::{NatsJetStreamSource, NatsPublisher};
 use super::{
     run_source, AsyncMessagePublisher, Bus, BusConsumer, MessageRouter, RunOptions, TransportError,
 };
-use crate::microsvc::{Message, MessageKind};
+use super::{Message, MessageKind};
 
 const DEFAULT_FETCH_TIMEOUT: Duration = Duration::from_millis(500);
